backend/requirements.txt
langchain
langchain-openai
langchain-community
flask
flask-cors
gevent
azure-cognitiveservices-speech
pypdf
faiss-cpu